Feat: Update to last API version Refact: Simplification of example app
