+++ example-app/build.gradle
@@ -45,10 +45,6 @@ dependencies {
     //Keyple Common
     implementation project(path: ':keyple-plugins:android-nfc')
     implementation "org.eclipse.keyple:keyple-java-service:2.0.0-SNAPSHOT"
-
-    //For Calypso Extension examples
-    implementation "org.eclipse.keyple:keyple-java-card-calypso:2.0.0-SNAPSHOT"
-
     //For Generic Extension examples
     implementation "org.eclipse.keyple:keyple-java-card-generic:2.0.0-SNAPSHOT"
