+++ example-app/src/main/AndroidManifest.xml
@@ -19,18 +19,13 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".activity.MainActivity">
+        <activity android:name=".activity.CoreExamplesActivity"
+            android:screenOrientation="portrait" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity android:name=".activity.CoreExamplesActivity"
-            android:screenOrientation="portrait" />
-        <activity
-            android:name=".activity.CalypsoExamplesActivity"
-            android:screenOrientation="portrait">
             <!-- Required for NFC KEYPLE -->
             <intent-filter>
                 <action android:name="android.nfc.action.TECH_DISCOVERED" />
