+++ example-app/src/main/kotlin/org/eclipse/keyple/plugin/android/nfc/example/activity/AbstractExampleActivity.kt
@@ -22,24 +22,16 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.navigation.NavigationView
 import java.io.IOException
-import kotlinx.android.synthetic.main.activity_calypso_examples.drawerLayout
-import kotlinx.android.synthetic.main.activity_calypso_examples.eventRecyclerView
-import kotlinx.android.synthetic.main.activity_calypso_examples.navigationView
-import kotlinx.android.synthetic.main.activity_calypso_examples.toolbar
-import org.eclipse.keyple.core.service.ObservableReader
-import org.eclipse.keyple.core.service.Reader
+import kotlinx.android.synthetic.main.activity_core_examples.drawerLayout
+import kotlinx.android.synthetic.main.activity_core_examples.eventRecyclerView
+import kotlinx.android.synthetic.main.activity_core_examples.navigationView
+import kotlinx.android.synthetic.main.activity_core_examples.toolbar
 import org.eclipse.keyple.core.service.ReaderEvent
-import org.eclipse.keyple.core.service.SmartCardServiceProvider
 import org.eclipse.keyple.core.service.selection.CardSelectionService
-import org.eclipse.keyple.core.service.selection.spi.SmartCard
 import org.eclipse.keyple.core.service.spi.ReaderObservationExceptionHandlerSpi
 import org.eclipse.keyple.core.service.spi.ReaderObserverSpi
-import org.eclipse.keyple.core.util.ByteArrayUtil
-import org.eclipse.keyple.core.util.protocol.ContactlessCardCommonProtocol
-import org.eclipse.keyple.plugin.android.nfc.AndroidNfcPluginFactoryAdapter
 import org.eclipse.keyple.plugin.android.nfc.example.R
 import org.eclipse.keyple.plugin.android.nfc.example.adapter.EventAdapter
-import org.eclipse.keyple.plugin.android.nfc.example.model.ChoiceEventModel
 import org.eclipse.keyple.plugin.android.nfc.example.model.EventModel
 import timber.log.Timber
 
@@ -58,8 +50,6 @@ abstract class AbstractExampleActivity : AppCompatActivity(), NavigationView.OnN
     private lateinit var adapter: RecyclerView.Adapter<*>
     private lateinit var layoutManager: RecyclerView.LayoutManager
     protected val events = arrayListOf<EventModel>()
-
-    protected lateinit var reader: Reader
 
     protected var useCase: UseCase? = null
     protected lateinit var cardSelectionsService: CardSelectionService
@@ -89,22 +79,6 @@ abstract class AbstractExampleActivity : AppCompatActivity(), NavigationView.OnN
         drawerLayout.addDrawerListener(toggle)
         toggle.syncState()
 
-        /**
-         * Register AndroidNfc plugin Factory
-         */
-        val plugin = SmartCardServiceProvider.getService().registerPlugin(AndroidNfcPluginFactoryAdapter(this))
-
-        /**
-         * Configure Nfc Reader
-         */
-        with(plugin.readers.values.first() as ObservableReader) {
-            setReaderObservationExceptionHandler(this@AbstractExampleActivity)
-            addObserver(this@AbstractExampleActivity)
-
-            // with this protocol settings we activate the nfc for ISO1443_4 protocol
-            activateProtocol(ContactlessCardCommonProtocol.ISO_14443_4.name, ContactlessCardCommonProtocol.ISO_14443_4.name)
-            reader = this
-        }
         window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
     }
 
@@ -150,12 +124,6 @@ abstract class AbstractExampleActivity : AppCompatActivity(), NavigationView.OnN
         Timber.d("Result: %s", message)
     }
 
-    protected fun addChoiceEvent(title: String, choices: List<String>, callback: (choice: String) -> Unit) {
-        events.add(ChoiceEventModel(title, choices, callback))
-        adapter.notifyItemInserted(events.lastIndex)
-        Timber.d("Choice: %s: %s", title, choices.toString())
-    }
-
     @Throws(IOException::class)
     protected fun checkNfcAvailability() {
         val nfcAdapter = NfcAdapter.getDefaultAdapter(this)
@@ -167,33 +135,6 @@ abstract class AbstractExampleActivity : AppCompatActivity(), NavigationView.OnN
                 throw IOException("Please enable NFC to communicate with NFC Elements\"")
             }
         }
-    }
-
-    override fun onDestroy() {
-        SmartCardServiceProvider.getService().plugins.forEach {
-            SmartCardServiceProvider.getService().unregisterPlugin(it.key)
-        }
-        super.onDestroy()
-    }
-
-    protected fun getSmardCardInfos(smartCard: SmartCard, index: Int): String {
-        val atr = try {
-            ByteArrayUtil.toHex(smartCard.atrBytes)
-        } catch (e: IllegalStateException) {
-            Timber.w(e)
-            e.message
-        }
-        val fci = try {
-            ByteArrayUtil.toHex(smartCard.fciBytes)
-        } catch (e: IllegalStateException) {
-            Timber.w(e)
-            e.message
-        }
-
-        return "Selection status for selection " +
-                "(indexed $index): \n\t\t" +
-                "ATR: ${atr}\n\t\t" +
-                "FCI: $fci"
     }
 
     override fun onReaderObservationError(pluginName: String?, readerName: String?, e: Throwable?) {
