+++ example-app/src/main/res/menu/core_examples_menu.xml
@@ -2,27 +2,9 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android" >
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/usecase1"
-            android:icon="@drawable/ic_nfc_black_24dp"
-            android:title="UC1: Explicit Selection"
-            android:checked="true"
-            />
-        <item
             android:id="@+id/usecase2"
             android:icon="@drawable/ic_nfc_black_24dp"
-            android:title="UC2: Default selection notification"
-            android:checked="true"
-            />
-        <item
-            android:id="@+id/usecase3"
-            android:icon="@drawable/ic_nfc_black_24dp"
-            android:title="UC3: Grouped Multi Selection"
-            android:checked="true"
-            />
-        <item
-            android:id="@+id/usecase4"
-            android:icon="@drawable/ic_nfc_black_24dp"
-            android:title="UC4: Sequential Multi Selection"
+            android:title="Default selection"
             android:checked="true"
             />
     </group>
